feat: ✨ folder structure change
